requirements.txt
appdirs==1.4.4
blinker==1.8.1
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
fs==2.4.16
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
Werkzeug==3.0.2
